feat: 將新增列表的部分拆到子元件，並且傳遞board_id到該component中
